Add documentation and wiring guide for STMBOY project
- Created BUTTON_WIRING_GUIDE.md detailing button connections on breadboard with STM32F411RE.
- Added TECHNICAL_OVERVIEW.md summarizing project architecture and system layers.
- Introduced TFM_DOCUMENTACION_TECNICA.md as a comprehensive technical document for the final project report.

+++ lib/platform/nucleof411re/platform_nucleof411re_serial.c
@@ -55,7 +55,7 @@ int _write(int file, char *ptr, int len)
 {
     (void)file;
 
-    for (uint8_t i = 0; i < len; i++)
+    for (int i = 0; i < len; i++)
     {
         const uint8_t byte = (uint8_t)ptr[i];
         send_byte_blocking(byte);
